+++ native/cocos/audio/common/decoder/AudioDecoder.h
@@ -116,7 +116,6 @@ public:
     virtual AudioDataFormat getDataFormat() const;
 
     virtual PCMHeader getPCMHeader() const;
-
 protected:
     AudioDecoder();
     virtual ~AudioDecoder();
