+++ native/cocos/audio/common/decoder/AudioDecoderManager.h
