+++ native/cocos/audio/common/utils/minifloat.cpp
@@ -15,7 +15,7 @@
  */
 
 #include <cmath>
-#include "audio/android/audio_utils/include/audio_utils/minifloat.h"
+#include "audio/common/utils/include/minifloat.h"
 
 #define EXPONENT_BITS 3
 #define EXPONENT_MAX  ((1 << EXPONENT_BITS) - 1)
