+++ native/cocos/audio/common/utils/primitives.cpp
@@ -14,20 +14,26 @@
  * limitations under the License.
  */
 
-#include "audio/android/audio_utils/include/audio_utils/primitives.h"
-#include "audio/android/audio_utils/private/private.h"
-#include "audio/android/cutils/bitops.h" /* for popcount() */
+#include "audio/common/utils/include/primitives.h"
+#include "audio/common/utils/private/private.h"
+#if CC_PLATFORM == CC_PLATFORM_ANDROID
+    #include "audio/android/cutils/bitops.h" /* for popcount() */
+#else
+    #include <cocos/base/Utils.h>
+    using namespace cc::utils;
+#endif
 
+//namespace {
 void ditherAndClamp(int32_t *out, const int32_t *sums, size_t c) {
     size_t i;
     for (i = 0; i < c; i++) {
-        int32_t l  = *sums++;
-        int32_t r  = *sums++;
+        int32_t l = *sums++;
+        int32_t r = *sums++;
         int32_t nl = l >> 12;
         int32_t nr = r >> 12;
-        l          = clamp16(nl);
-        r          = clamp16(nr);
-        *out++     = (r << 16) | (l & 0xFFFF);
+        l = clamp16(nl);
+        r = clamp16(nr);
+        *out++ = (r << 16) | (l & 0xFFFF);
     }
 }
 
@@ -241,8 +247,8 @@ void downmix_to_mono_i16_from_stereo_i16(int16_t *dst, const int16_t *src, size_
 void upmix_to_stereo_i16_from_mono_i16(int16_t *dst, const int16_t *src, size_t count) {
     while (count--) {
         int32_t temp = *src++;
-        dst[0]       = temp;
-        dst[1]       = temp;
+        dst[0] = temp;
+        dst[1] = temp;
         dst += 2;
     }
 }
@@ -257,8 +263,8 @@ void downmix_to_mono_float_from_stereo_float(float *dst, const float *src, size_
 void upmix_to_stereo_float_from_mono_float(float *dst, const float *src, size_t frames) {
     while (frames--) {
         float temp = *src++;
-        dst[0]     = temp;
-        dst[1]     = temp;
+        dst[0] = temp;
+        dst[1] = temp;
         dst += 2;
     }
 }
@@ -343,21 +349,21 @@ void memcpy_by_channel_mask(void *dst, uint32_t dst_mask,
     }
     switch (sample_size) {
         case 1: {
-            uint8_t *udst = (uint8_t *)dst;
-            const uint8_t *usrc = (const uint8_t *)src;
+            auto *udst = static_cast<uint8_t *>(dst);
+            const auto *usrc = static_cast<const uint8_t *>(src);
 
             copy_frame_by_mask(udst, dst_mask, usrc, src_mask, count, 0);
         } break;
         case 2: {
-            uint16_t *udst = (uint16_t *)dst;
-            const uint16_t *usrc = (const uint16_t *)src;
+            auto *udst = static_cast<uint16_t *>(dst);
+            const auto *usrc = (const uint16_t *)src;
 
             copy_frame_by_mask(udst, dst_mask, usrc, src_mask, count, 0);
         } break;
         case 3: { /* could be slow.  use a struct to represent 3 bytes of data. */
             uint8x3_t *udst = (uint8x3_t *)dst;
             const uint8x3_t *usrc = (const uint8x3_t *)src;
-            static const uint8x3_t zero; /* tricky - we use this to zero out a sample */
+            static const uint8x3_t zero{0,0,0}; /* tricky - we use this to zero out a sample */
 
             copy_frame_by_mask(udst, dst_mask, usrc, src_mask, count, zero);
         } break;
@@ -382,44 +388,44 @@ void memcpy_by_channel_mask(void *dst, uint32_t dst_mask,
 #define copy_frame_by_idx(dst, dst_channels, src, src_channels, idxary, count, zero) \
     {                                                                                \
         unsigned i;                                                                  \
-        int      index;                                                              \
+        int index;                                                                   \
         while ((count)--) {                                                          \
             for (i = 0; i < (dst_channels); ++i) {                                   \
-                index    = (idxary)[i];                                              \
+                index = (idxary)[i];                                                 \
                 *(dst)++ = index < 0 ? (zero) : (src)[index];                        \
             }                                                                        \
             (src) += (src_channels);                                                 \
         }                                                                            \
     }
 
-void memcpy_by_index_array(void *dst, uint32_t dst_channels,
-                           const void *src, uint32_t src_channels,
-                           const int8_t *idxary, size_t sample_size, size_t count) {
-    switch (sample_size) {
+void memcpy_by_index_array(void *dst, uint32_t dstChannels,
+                           const void *src, uint32_t srcChannels,
+                           const int8_t *idxary, size_t sampleSize, size_t count) {
+    switch (sampleSize) {
         case 1: {
-            uint8_t *      udst = (uint8_t *)dst;
-            const uint8_t *usrc = (const uint8_t *)src;
+            auto *udst = static_cast<uint8_t *>(dst);
+            const auto *usrc = static_cast<const uint8_t *>(src);
 
-            copy_frame_by_idx(udst, dst_channels, usrc, src_channels, idxary, count, 0);
+            copy_frame_by_idx(udst, dstChannels, usrc, srcChannels, idxary, count, 0); // NOLINT
         } break;
         case 2: {
-            uint16_t *      udst = (uint16_t *)dst;
-            const uint16_t *usrc = (const uint16_t *)src;
+            auto *udst = static_cast<uint16_t *>(dst);
+            const auto *usrc = static_cast<const uint16_t *>(src);
 
-            copy_frame_by_idx(udst, dst_channels, usrc, src_channels, idxary, count, 0);
+            copy_frame_by_idx(udst, dstChannels, usrc, srcChannels, idxary, count, 0); // NOLINT
         } break;
         case 3: { /* could be slow.  use a struct to represent 3 bytes of data. */
-            uint8x3_t *            udst = (uint8x3_t *)dst;
-            const uint8x3_t *      usrc = (const uint8x3_t *)src;
-            static const uint8x3_t zero;
+            auto *udst = (uint8x3_t *)dst;
+            const uint8x3_t *usrc = (const uint8x3_t *)src;
+            static const uint8x3_t zero{0,0,0};
 
-            copy_frame_by_idx(udst, dst_channels, usrc, src_channels, idxary, count, zero);
+            copy_frame_by_idx(udst, dstChannels, usrc, srcChannels, idxary, count, zero);
         } break;
         case 4: {
-            uint32_t *      udst = (uint32_t *)dst;
+            uint32_t *udst = (uint32_t *)dst;
             const uint32_t *usrc = (const uint32_t *)src;
 
-            copy_frame_by_idx(udst, dst_channels, usrc, src_channels, idxary, count, 0);
+            copy_frame_by_idx(udst, dstChannels, usrc, srcChannels, idxary, count, 0);
         } break;
         default:
             abort(); /* illegal value */
@@ -429,8 +435,8 @@ void memcpy_by_index_array(void *dst, uint32_t dst_channels,
 
 size_t memcpy_by_index_array_initialization(int8_t *idxary, size_t idxcount,
                                             uint32_t dst_mask, uint32_t src_mask) {
-    size_t   n      = 0;
-    int      srcidx = 0;
+    size_t n = 0;
+    int srcidx = 0;
     uint32_t bit, ormask = src_mask | dst_mask;
 
     while (ormask && n < idxcount) {
@@ -472,8 +478,8 @@ size_t memcpy_by_index_array_initialization_src_index(int8_t *idxary, size_t idx
 size_t memcpy_by_index_array_initialization_dst_index(int8_t *idxary, size_t idxcount,
                                                       uint32_t dst_mask, uint32_t src_mask) {
     size_t src_idx, dst_idx;
-    size_t dst_count = __builtin_popcount(dst_mask);
-    size_t src_count = __builtin_popcount(src_mask);
+    size_t dst_count = popcount(dst_mask);
+    size_t src_count = popcount(src_mask);
     if (idxcount == 0) {
         return dst_count;
     }
@@ -488,3 +494,4 @@ size_t memcpy_by_index_array_initialization_dst_index(int8_t *idxary, size_t idx
     }
     return dst_idx;
 }
+//} // namespace
